Organizes Locations, Quests, and Tasks
Adds /Location, /Quest, /Task
