+++ src/scripts/components/Quest/Card.js
@@ -1,49 +1,47 @@
 'use strict';
 
 var React = require('react/addons');
-var Markdown = require('markdown').markdown;
 
 var ButtonToolbar = require('react-bootstrap').ButtonToolbar;
 var Button = require('react-bootstrap').Button;
 var ButtonLink = require('react-router-bootstrap').ButtonLink;
 var Glyphicon = require('react-bootstrap').Glyphicon;
 
-var FormModal = require('./TaskFormModal');
-var RemoveModal = require('./TaskRemoveModal');
-var AttrBlock = require('./ModelAttrBlock');
+var FormModal = require('./FormModal');
+var RemoveModal = require('./RemoveModal');
+var AttrBlock = require('../ModelAttrBlock');
 
-require('../../styles/ItemCard.css');
+require('../../../styles/ItemCard.css');
 
 
 
-var TaskCard = React.createClass({
+var QuestCard = React.createClass({
     render: function () {
         var campaign = this.props.campaign;
         var act = this.props.act;
         var quest = this.props.quest;
-        var task = this.props.task;
 
         return (
-            <div className="item-card task-card">
+            <div className="item-card quest-card">
                 <div className="card-header">
                     <p className="pull-left">
-                        {task.attrs.name}
+                        {quest.attrs.name}
                     </p>
                     <p className="pull-right">
-                        <span />
+                        <small className="text-muted">{quest.attrs.type}&nbsp;&nbsp;<strong>CR {quest.attrs.challenge}</strong></small>
                     </p>
                 </div>
                 <div className="card-body">
-                    <AttrBlock name="Flavor" attr={task.attrs.flavor} markdown />
-                    <AttrBlock name="Details" attr={task.attrs.details} markdown />
+                    <AttrBlock name="Flavor" attr={quest.attrs.flavor} markdown />
+                    <AttrBlock name="Rewards" attr={quest.attrs.rewardOther} markdown />
                 </div>
                 <div className="card-footer">
                     <ButtonToolbar className="pull-left">
-                        <Button bsStyle="warning" bsSize="small"><Glyphicon glyph="plus" /> Encounter</Button>
+                        <ButtonLink bsStyle="primary" bsSize="small" to="manage-tasks" params={{campaignId: campaign.id, actId: act.id, questId: quest.id}}><Glyphicon glyph="cog" /> Tasks</ButtonLink>
                     </ButtonToolbar>
                     <ButtonToolbar className="pull-right">
-                        <FormModal quest={quest} task={task} onUpdate={this.props.onUpdate} />
-                        <RemoveModal task={task} onUpdate={this.props.onUpdate} />
+                        <FormModal act={act} quest={quest} onUpdate={this.props.onUpdate} />
+                        <RemoveModal quest={quest} onUpdate={this.props.onUpdate} />
                     </ButtonToolbar>
                 </div>
             </div>
@@ -51,6 +49,6 @@ var TaskCard = React.createClass({
     }
 });
 
-module.exports = TaskCard;
+module.exports = QuestCard;
