+++ src/scripts/components/Quest/ManagePage.js
@@ -3,15 +3,15 @@
 var React = require('react/addons');
 var Router = require('react-router');
 
-var Auth = require('../helpers/Auth');
+var Auth = require('../../helpers/Auth');
 
-var CampaignModel = require('../models/CampaignModel');
-var ActModel = require('../models/ActModel');
+var CampaignModel = require('../../models/CampaignModel');
+var ActModel = require('../../models/ActModel');
 
-var Breadcrumb = require('./MainBreadcrumb');
-var PageHeader = require('./ModelPageHeader');
-var FormModal = require('./QuestFormModal');
-var Card = require('./QuestCard');
+var Breadcrumb = require('../MainBreadcrumb');
+var PageHeader = require('../ModelPageHeader');
+var FormModal = require('./FormModal');
+var Card = require('./Card');
